Initializable Temporal clients
